Add inventory-service

+++ Order/src/main/java/com/example/order/repository/CartItemRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartItemRepository extends JpaRepository<CartItem,Integer> {
-    void deleteCartItemByProductID(Integer id);
+public interface CartItemRepository extends JpaRepository<CartItem,Long> {
+    void deleteCartItemByProductID(Long id);
 }
