+++ Order/src/main/java/com/example/order/repository/OrderRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order,Integer>, JpaSpecificationExecutor<Order> {
+public interface OrderRepository extends JpaRepository<Order,Long>, JpaSpecificationExecutor<Order> {
 }
